* Moved tests to Node
